Session06/assignment/app-pydantic/pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.51.0",
    "mcp[cli]>=1.22.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pydantic-core>=2.41.5",
    "rich>=14.2.0",
]
